+++ api/utils/querydb.py
@@ -4,6 +4,7 @@ from sqlmodel import select
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from models.models import (
+    NPC,
     BossPals,
     Breeding,
     BuidObjects,
@@ -12,18 +13,114 @@ from models.models import (
     Gear,
     Items,
     Pals,
+    PassiveSkills,
     SickPal,
     TechTree,
-    PassiveSkills,
-    NPC,
 )
 
 
-async def get_item(db: AsyncSession, name: str, variety: str):
-    if name:
-        return await paginate(db, select(Items).where(Items.Name == name))
-    if variety:
-        return await paginate(db, select(Items).where(Items.Type == variety))
+async def get_pals(db: AsyncSession, params):
+    palstocolume = {
+        "name": Pals.Name,
+        "dexkey": Pals.DexKey,
+        "nocturnal": Pals.Nocturnal,
+    }
+    wheres = select(Pals)
+    for parm in params:
+        if parm == "size" or parm == "page":
+            continue
+        if parm == "nocturnal":
+            if params[parm] == "True":
+                p = 1
+            else:
+                p = 0
+        else:
+            p = params[parm]
+        if parm == "type":
+            wheres = wheres.where(
+                text(
+                    "JSON_SEARCH(Types, 'one', :name COLLATE utf8mb4_general_ci)"
+                ).bindparams(name=p)
+            )
+        elif parm == "suitability":
+            wheres = wheres.where(
+                text(
+                    "JSON_SEARCH(Suitability, 'one', :suit COLLATE utf8mb4_general_ci)"
+                ).bindparams(suit=p)
+            )
+        elif parm == "drop":
+            wheres = wheres.where(
+                text(
+                    f"JSON_SEARCH(Drops, 'one', :drop COLLATE utf8mb4_general_ci)"
+                ).bindparams(drop=p)
+            )
+        elif parm == "skill":
+            wheres = wheres.where(
+                text(
+                    "JSON_SEARCH(Skills, 'one', :skill COLLATE utf8mb4_general_ci)"
+                ).bindparams(skill=p)
+            )
+        else:
+            wheres = wheres.where(palstocolume[parm] == p)
+    return await paginate(db, wheres)
+
+
+async def get_bosspal(db: AsyncSession, params):
+    palstocolume = {
+        "name": BossPals.Name,
+        "nocturnal": BossPals.Nocturnal,
+    }
+    wheres = select(BossPals)
+    for parm in params:
+        if parm == "size" or parm == "page":
+            continue
+        if parm == "nocturnal":
+            if params[parm] == "True":
+                p = 1
+            else:
+                p = 0
+        else:
+            p = params[parm]
+        if parm == "type":
+            wheres = wheres.where(
+                text(
+                    "JSON_SEARCH(Types, 'one', :name COLLATE utf8mb4_general_ci)"
+                ).bindparams(name=p)
+            )
+        elif parm == "suitability":
+            wheres = wheres.where(
+                text(
+                    "JSON_SEARCH(Suitability, 'one', :suit COLLATE utf8mb4_general_ci)"
+                ).bindparams(suit=p)
+            )
+        elif parm == "drop":
+            wheres = wheres.where(
+                text(
+                    f"JSON_SEARCH(Drops, 'one', :drop COLLATE utf8mb4_general_ci)"
+                ).bindparams(drop=p)
+            )
+        elif parm == "skill":
+            wheres = wheres.where(
+                text(
+                    "JSON_SEARCH(Skills, 'one', :skill COLLATE utf8mb4_general_ci)"
+                ).bindparams(skill=p)
+            )
+        else:
+            wheres = wheres.where(palstocolume[parm] == p)
+    return await paginate(db, wheres)
+
+
+async def get_item(db: AsyncSession, params):
+    itemtocolume = {
+        "name": Items.Name,
+        "type": Items.Type,
+    }
+    wheres = select(Items)
+    for parm in params:
+        if parm == "size" or parm == "page":
+            continue
+        wheres = wheres.where(itemtocolume[parm] == params[parm])
+    return await paginate(db, wheres)
 
 
 async def get_crafting(db: AsyncSession, name: str):
@@ -32,77 +129,6 @@ async def get_crafting(db: AsyncSession, name: str):
 
 async def get_gear(db: AsyncSession, name: str):
     return await paginate(db, select(Gear).where(Gear.Name == name))
-
-
-async def get_pal(db: AsyncSession, name: str):
-    return await paginate(db, select(Pals).where(Pals.Name == name))
-
-
-async def get_bosspal(db: AsyncSession, name: str):
-    return await paginate(db, select(BossPals).where(BossPals.Name == name))
-
-
-async def get_pal_by_dexid(db: AsyncSession, idx: str):
-    return await paginate(db, select(Pals).where(Pals.DexKey == idx))
-
-
-async def get_pal_by_type(db: AsyncSession, name: str):
-    statement = select(Pals).where(
-        text("JSON_SEARCH(Types, 'one', :name COLLATE utf8mb4_general_ci)").bindparams(
-            name=name
-        )
-    )
-    return await paginate(db, statement)
-
-
-async def get_bosspal_by_type(db: AsyncSession, name: str):
-    statement = select(BossPals).where(
-        text("JSON_SEARCH(Types, 'one', :name COLLATE utf8mb4_general_ci)").bindparams(
-            name=name
-        )
-    )
-    return await paginate(db, statement)
-
-
-async def get_pal_by_suitability(db: AsyncSession, name: str):
-    statement = select(Pals).where(
-        text(
-            "JSON_SEARCH(Suitability, 'one', :name COLLATE utf8mb4_general_ci)"
-        ).bindparams(name=name)
-    )
-    return await paginate(db, statement)
-
-
-async def get_bosspal_by_suitability(db: AsyncSession, name: str):
-    statement = select(BossPals).where(
-        text(
-            "JSON_SEARCH(Suitability, 'one', :name COLLATE utf8mb4_general_ci)"
-        ).bindparams(name=name)
-    )
-    return await paginate(db, statement)
-
-
-async def get_pal_by_drops(db: AsyncSession, drop: str):
-    statement = select(Pals).where(
-        text(f"JSON_SEARCH(Drops, 'one', :name COLLATE utf8mb4_general_ci)").bindparams(
-            name=drop
-        )
-    )
-    return await paginate(db, statement)
-
-
-async def get_pal_by_skills(db: AsyncSession, skill: str):
-    statement = select(Pals).where(
-        text("JSON_SEARCH(Skills, 'one', :name COLLATE utf8mb4_general_ci)").bindparams(
-            name=skill
-        )
-    )
-    return await paginate(db, statement)
-
-
-async def get_pal_by_nocturnal(db: AsyncSession, nocturnal: str):
-    statement = select(Pals).where(Pals.Nocturnal == nocturnal)
-    return await paginate(db, statement)
 
 
 async def get_foodeffects(db: AsyncSession, name: str):
@@ -134,11 +160,14 @@ async def get_build_by_category(db: AsyncSession, category: str):
         db, select(BuidObjects).where(BuidObjects.Category == category)
     )
 
+
 async def get_passive(db: AsyncSession, name: str):
     return await paginate(db, select(PassiveSkills).where(PassiveSkills.Name == name))
 
+
 async def get_npc(db: AsyncSession, name: str):
     return await paginate(db, select(NPC).where(NPC.Name == name))
+
 
 async def get_all(db: AsyncSession, name):
     if name == "pals":